Fixed text converter. Improved API error handling.

+++ pages/api/auth/logout.js
@@ -2,22 +2,34 @@ import { connection } from '@/src/db/connection';
 import { serialize } from 'cookie';
 
 const handler = async (req, res) => {
-    let token = req.cookies["session-token"];
+    let result = {
+        content: undefined,
+        errors: []
+    };
+    let status = 200;
 
-    await connection.query('DELETE FROM Session WHERE UserID = ?', [token]);
+    if (req.method === "GET") {
+        let token = req.cookies["session-token"];
 
-    await connection.end();
+        await connection.query('DELETE FROM Session WHERE UserID = ?', [token]);
 
-    const cookie = serialize("session-token", "DELETED", {
-        httpOnly: true,
-        path: "/",
-        sameSite: "strict",
-        secure: true
-    });
+        await connection.end();
 
-    res.setHeader("Set-Cookie", cookie);
+        const cookie = serialize("session-token", "DELETED", {
+            httpOnly: true,
+            path: "/",
+            sameSite: "strict",
+            secure: true
+        });
 
-    res.status(200).send();
+        res.setHeader("Set-Cookie", cookie);
+    }
+    else {
+        status = 405; // Method Not Allowed
+        result.errors.push("HTTP " + req.method + " is not a valid method for this API.");
+    }
+    
+    res.status(status).json(result);
 }
 
 export default handler;
